requirements.txt
passlib[bcrypt]
pyjwt
python-multipart
fastapi
uvicorn
pillow
ffmpeg-python
databases[sqlite]
sqlalchemy
cairosvg